reorganized initializers and added some macros
